refactor: changed the name of the module to engine
