+++ src/lib.rs
@@ -1,2 +1,2 @@
-pub mod run;
+pub mod engine;
 pub mod run_statistics;
